feat(dashboards/analysis): delegate data access to the service
